+++ Plugins/Core/Common/ListItem.cs
@@ -7,6 +7,7 @@
 	/// </summary>
 	public class ListItem
 	{
+
 		/// <summary>
 		/// The unique identifier or name for the list item.
 		/// </summary>